Add QuotaManager
